Move converter class
Also improve formatting.
