+++ src/main/kotlin/nl/avisi/structurizr/site/generatr/site/model/ContentText.kt
@@ -6,14 +6,8 @@ import com.structurizr.documentation.Section
 import com.vladsch.flexmark.ast.Heading
 import com.vladsch.flexmark.ast.Paragraph
 import com.vladsch.flexmark.parser.Parser
-import org.asciidoctor.Asciidoctor
 import org.asciidoctor.Options
 import org.asciidoctor.SafeMode
-import org.asciidoctor.ast.ContentNode
-import org.asciidoctor.ast.Document
-import org.asciidoctor.ast.StructuralNode
-import org.asciidoctor.converter.ConverterFor
-import org.asciidoctor.converter.StringConverter
 
 private val parser = Parser.builder().build()
 
@@ -54,22 +48,4 @@ private fun asciidocText(content: String): String {
     val text = asciidoctor.convert(content, options)
 
     return text.lines().joinToString(" ")
-}
-
-@ConverterFor("text")
-class AsciiDocTextConverter(
-    backend: String?,
-    opts: Map<String?, Any?>?
-) : StringConverter(backend, opts) {
-    override fun convert(node: ContentNode, transform: String?, o: Map<Any?, Any?>?): String? {
-        val transform1 = transform ?: node.nodeName
-        return if (node is Document)
-            node.content.toString()
-        else if (node is org.asciidoctor.ast.Section)
-            "${node.title}\n${node.content}"
-        else if (transform1 == "preamble" || transform1 == "paragraph")
-            (node as StructuralNode).content as String
-        else
-            null
-    }
 }
